cleaned up folders
