+++ frontend/angular.json
@@ -3,7 +3,9 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "p2-Angular": {
+    
+    "p2-beta": {
+      
       "projectType": "application",
       "schematics": {
         "@schematics/angular:application": {
@@ -17,14 +19,19 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "outputPath": "dist/p2-angular",
+            "outputPath": "dist/p2-beta",
             "index": "src/index.html",
             "main": "src/main.ts",
             "polyfills": "src/polyfills.ts",
             "tsConfig": "tsconfig.app.json",
             "assets": [
               "src/favicon.ico",
-              "src/assets"
+              "src/assets",
+              {
+                "glob": "**/*",
+                "input": "node_modules/leaflet/dist/images/", 
+                "output": "./assets/"
+             }
             ],
             "styles": [
               "src/styles.css"
@@ -68,10 +75,10 @@
           "builder": "@angular-devkit/build-angular:dev-server",
           "configurations": {
             "production": {
-              "browserTarget": "p2-Angular:build:production"
+              "browserTarget": "p2-beta:build:production"
             },
             "development": {
-              "browserTarget": "p2-Angular:build:development"
+              "browserTarget": "p2-beta:build:development"
             }
           },
           "defaultConfiguration": "development"
@@ -79,7 +86,7 @@
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "p2-Angular:build"
+            "browserTarget": "p2-beta:build"
           }
         },
         "test": {
@@ -102,5 +109,5 @@
       }
     }
   },
-  "defaultProject": "p2-Angular"
+  "defaultProject": "p2-beta"
 }
