+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "p2-angular",
+  "name": "p2-beta",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -18,6 +18,8 @@
     "@angular/platform-browser": "~13.3.0",
     "@angular/platform-browser-dynamic": "~13.3.0",
     "@angular/router": "~13.3.0",
+    "@asymmetrik/ngx-leaflet": "^13.0.2",
+    "leaflet": "^1.7.1",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
@@ -26,7 +28,9 @@
     "@angular-devkit/build-angular": "~13.3.0",
     "@angular/cli": "~13.3.0",
     "@angular/compiler-cli": "~13.3.0",
+    "@types/events": "^3.0.0",
     "@types/jasmine": "~3.10.0",
+    "@types/leaflet": "^1.7.9",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
     "karma": "~6.3.0",
