+++ frontend/src/app/waypoints/waypoints.component.html
