+++ frontend/src/app/waypoints/waypoints.component.ts
