+++ frontend/src/app/waypt-detail/waypt-detail.component.html
