+++ frontend/src/app/waypt-detail/waypt-detail.component.spec.ts
