+++ frontend/src/assets/rick.png
