+++ frontend/src/styles.css
@@ -1,1 +1,7 @@
-/* You can add global styles to this file, and also import other style files */
+@import "../node_modules/leaflet/dist/leaflet.css";
+
+#map { 
+    height: 300px;
+    width: 100vh; 
+    margin: auto;
+}
